+++ Headliner/Headliner/Source/Presentation/General/Path.swift
@@ -6,6 +6,14 @@
 //
 
 import Foundation
+import ShazamKit
+
+struct MediaRoute: Hashable {
+    let title: String
+    let artist: String
+    let artworkURL: URL?
+    let mediaItem: SHMediaItem?
+}
 
 enum PathType: Hashable {
     case loading
