+++ Headliner/Headliner/Source/Presentation/Shazam/ShazamSearchView.swift
@@ -2,12 +2,7 @@ import SwiftUI
 import ShazamKit
 import MusicKit
 
-struct MediaRoute: Hashable {
-    let title: String
-    let artist: String
-    let artworkURL: URL?
-    let mediaItem: SHMediaItem?
-}
+
 
 struct ShazamSearchView: View {
     @EnvironmentObject var pathModel: PathModel
@@ -31,20 +26,7 @@ struct ShazamSearchView: View {
                         LazyVStack(spacing: 0) {
                             ForEach(viewModel.results) { song in
                                 Button {
-                                    let properties: [SHMediaItemProperty: Any] = [
-                                        .title: song.title,
-                                        .artist: song.artistName,
-                                        .artworkURL: song.artworkURL as Any
-                                    ]
-                                    let mediaItem = SHMediaItem(properties: properties)
-                                    
-                                    let route = MediaRoute(
-                                        title: song.title,
-                                        artist: song.artistName,
-                                        artworkURL: song.artworkURL,
-                                        mediaItem: mediaItem
-                                    )
-                                    pathModel.paths.append(.result(route))
+                                    viewModel.handleMusicSelection(song: song)
                                 } label: {
                                     MusicRowView(
                                         title: song.title,
@@ -77,15 +59,9 @@ struct ShazamSearchView: View {
         .task {
             await viewModel.prepare()
         }
-        .onChange(of: viewModel.currentItem) { _, item in
-            if let item {
-                let route = MediaRoute(
-                    title: item.title ?? "",
-                    artist: item.artist ?? "",
-                    artworkURL: item.artworkURL,
-                    mediaItem: item
-                )
-                pathModel.paths.append(.result(route))
+        .onChange(of: viewModel.navigationRoute) { _, route in
+            if let route = route {
+                pathModel.paths.append(route)
             }
         }
     }
@@ -94,7 +70,6 @@ struct ShazamSearchView: View {
         Button{
             if !viewModel.isListening {
                 viewModel.start()
-                pathModel.paths.append(.loading)
             }
         } label: {
             Image(.shazamButton)
